@@ -17,33 +17,33 @@
         "prepare": "husky install"
     },
     "devDependencies": {
-        "@commitlint/cli": "^12.1.4",
-        "@commitlint/config-conventional": "^12.1.4",
+        "@commitlint/cli": "^14.1.0",
+        "@commitlint/config-conventional": "^14.1.0",
         "@nomiclabs/hardhat-ethers": "^2.0.0",
-        "@nomiclabs/hardhat-etherscan": "^2.1.4",
+        "@nomiclabs/hardhat-etherscan": "^2.1.7",
         "@nomiclabs/hardhat-waffle": "^2.0.0",
-        "@typechain/ethers-v5": "^7.0.1",
-        "@typechain/hardhat": "^2.1.2",
-        "@types/chai": "^4.2.21",
-        "@types/mocha": "^8.2.3",
+        "@typechain/ethers-v5": "^8.0.2",
+        "@typechain/hardhat": "^3.0.0",
+        "@types/chai": "^4.2.22",
+        "@types/mocha": "^9.0.0",
         "babel-eslint": "^10.1.0",
         "chai": "^4.2.0",
         "cross-env": "^7.0.3",
         "dotenv": "^10.0.0",
-        "eslint": "^7.30.0",
+        "eslint": "^8.2.0",
         "ethereum-waffle": "^3.0.0",
-        "ethers": "^5.0.0",
-        "hardhat": "^2.4.3",
-        "hardhat-dependency-compiler": "^1.1.1",
+        "ethers": "^5.5.1",
+        "hardhat": "^2.6.8",
+        "hardhat-dependency-compiler": "^1.1.2",
         "hardhat-gas-reporter": "^1.0.4",
-        "husky": "^7.0.1",
-        "prettier": "^2.3.2",
-        "prettier-plugin-solidity": "^1.0.0-beta.14",
+        "husky": "^7.0.4",
+        "prettier": "^2.4.1",
+        "prettier-plugin-solidity": "^1.0.0-beta.19",
         "solhint-plugin-prettier": "^0.0.5",
-        "solidity-coverage": "^0.7.16",
+        "solidity-coverage": "^0.7.17",
         "ts-generator": "^0.1.1",
-        "ts-node": "^10.1.0",
-        "typechain": "^5.1.1",
-        "typescript": "^4.3.5"
+        "ts-node": "^10.4.0",
+        "typechain": "^6.0.2",
+        "typescript": "^4.4.4"
     }
 }
